REORG/MINOR: lua: convert boolean "int" to bitfield
This patch converts a boolean "int" to a bitfiled. The main
reason is to save space in the struct if another flag may will
be require.

Note that this patch is required for next fix and will need to be
backported to 1.6.
(cherry picked from commit 7fa0549a2b7eaedd1a8bf9e13266b6535a9257fb)