MINOR: lru: new function to delete <nb> least recently used keys
Introduction of a new function in the LRU cache source file.
Purpose of this function is to be used to delete a number of entries in
the cache. 'number' is defined by the caller and the key removed are
taken at the tail of the tree
(cherry picked from commit 22c4ed69372d677061fb1f400311c57a8d4ce2cf)

[wt: needed in 1.6 for the few subsequent 51d fixes]