MINOR: cfgparse: warn for incorrect 'timeout retry' keyword spelling in resolvers
If for example it was written as 'timeout retri 1s' or 'timeout wrong 1s'
this would be used for the retry timeout value. Resolvers section only
timeout setting currently is 'retry', others are still parsed as before
this patch to not break existing configurations.

This should only be applied as a backport to 1.6, for 1.7 a different
patch is provided to abort if a wrong keyword is used (commit 7a91a0e1).