CLEANUP: payload: remove useless and confusing nullity checks for channel buffer
The buffer could not be null by definition since we moved the stream out
of the session. It's the stream which gained the ability to be NULL (hence
the recent fix for this case). The checks in place are useless and make one
think this situation can happen.
(cherry picked from commit 4cefbc07529923d3d0e89fba52c2e037d70e664e)